requirements.txt
sanic
#torch==1.7.1
#torchvision==0.8.2
torch==1.12.1
torchvision==0.13.1
numpy==1.23.3
lmdb==1.2.1
#opencv-python==4.5.3.56
opencv-python==4.6.0.66
PyYAML==5.4.1
tqdm==4.62.2
yapf==0.31.0
basicsr==1.4.2
facexlib==0.2.5
gfpgan==1.3.8
Pillow
requests_futures